Simplify installation location retrieval (#18)
This replaces the whole existing logic to determine the path of an installed package with a new one.

Instead of a lot of guessing we now read the output of npm. Unfortunately we have to parse the human readable output since the information we need (which package was just installed) is not reliably extractable from the output generated when passing --json or --parseable.

Apart from a tremendous simplification of the existing code, with this change we now transparently support all spec formats that npm supports too.

The removal of getPath is a possibly breaking change. The function was neither documented nor used in any apache/cordova-* repositories though.

This supports all npm versions from 6 down to 3 (which are all that are bundled with supported node versions).