chore(ci): Fix dependabot code scanning errors
Dependabot pushes a branch to the repo and opens a pull request. The
pull request workflow runs properly, but the push workflow fails because
dependabot doesn't have access to a read/write GITHUB_TOKEN and doesn't
have access to the repo secrets.

The quick fix is telling CI not to run the push event workflow for
branches that start with `dependabot/`.